refactor: finished preprocessing_train, workon training
