+++ qdf/settings.py
@@ -1,0 +1,10 @@
+# Dataset.
+DATASET="QM9under7atoms_atomizationenergy_eV"
+# DATASET=QM9under14atoms_atomizationenergy_eV
+# DATASET=QM9full_atomizationenergy_eV
+# DATASET=QM9under7atoms_homolumo_eV  # Train homo and lumo simultaneously.
+# DATASET=QM9full_homolumo_eV
+# DATASET=yourdataset_property_unit
+
+DATASET_PATH = "/home/raul/git/QuantumDeepField_molecule/dataset"
+SAVE_PATH = "/home/raul/git/QuantumDeepField_molecule/output"
